+++ packages/x-components/src/x-modules/queries-preview/components/query-preview.vue
@@ -119,6 +119,8 @@
        */
       const queryPreviewHash = computed(() => getHashFromQueryPreviewInfo(props.queryPreviewInfo));
 
+      provide('queryPreviewHash', queryPreviewHash);
+
       /**
        * Gets from the state the results preview of the query preview.
        *